requirements.txt
# ===========================================
# 量化交易系统依赖清单
# ===========================================

# 数据源
baostock>=0.8.8          # 主要数据源（免费）
akshare>=1.10.0          # 备用数据源
tushare>=1.3.0           # 高级数据源（需要 Token）

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0          # Parquet 文件支持

# 配置文件
pyyaml>=6.0

# 网络请求
requests>=2.28.0

# 进度条
tqdm>=4.65.0

# 可选：GUI 依赖（按需安装）
# PyQt5>=5.15.0
# PyQtWebEngine>=5.15.0

# 可选：性能优化（按需安装）
# numba>=0.57.0          # JIT 加速

# 可选：回测可视化（按需安装）
# matplotlib>=3.7.0
# seaborn>=0.12.0

# 开发依赖（按需安装）
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# mypy>=1.0.0
